Try to update dependencies to resolve conflicts
