+++ src/Microsoft.DotNet.Darc/DarcLib/Helpers/DependencyFileManager.cs
@@ -313,6 +313,29 @@ public class DependencyFileManager : IDependencyFileManager
         JObject toolsConfigurationJson = await ReadDotNetToolsConfigJsonAsync(repoUri, branch, repoIsVmr);
         XmlDocument nugetConfig = await ReadNugetConfigAsync(repoUri, branch);
 
+        return UpdateDependencyFiles(
+            versionDetails,
+            versionProps,
+            globalJson,
+            toolsConfigurationJson,
+            nugetConfig,
+            itemsToUpdate,
+            sourceDependency,
+            oldDependencies,
+            incomingDotNetSdkVersion);
+    }
+
+    public GitFileContentContainer UpdateDependencyFiles(
+        XmlDocument versionDetails,
+        XmlDocument versionProps,
+        JObject globalJson,
+        JObject toolsConfigurationJson,
+        XmlDocument nugetConfig,
+        IEnumerable<DependencyDetail> itemsToUpdate,
+        SourceDependency sourceDependency,
+        IEnumerable<DependencyDetail> oldDependencies,
+        SemanticVersion incomingDotNetSdkVersion)
+    {
         foreach (DependencyDetail itemToUpdate in itemsToUpdate)
         {
             try
@@ -321,7 +344,7 @@ public class DependencyFileManager : IDependencyFileManager
             }
             catch (DarcException e)
             {
-                throw new DarcException(e.Message + $" in repo '{repoUri}' and branch '{branch}'", e);
+                throw new DarcException(e.Message, e);
             }
 
             UpdateVersionFiles(versionProps, globalJson, toolsConfigurationJson, itemToUpdate);
