+++ src/Microsoft.DotNet.Darc/DarcLib/Helpers/IDependencyFileManager.cs
@@ -58,6 +58,17 @@ public interface IDependencyFileManager
         IEnumerable<DependencyDetail> oldDependencies,
         SemanticVersion? incomingDotNetSdkVersion);
 
+    GitFileContentContainer UpdateDependencyFiles(
+        XmlDocument versionDetails,
+        XmlDocument versionProps,
+        JObject globalJson,
+        JObject toolsConfigurationJson,
+        XmlDocument nugetConfig,
+        IEnumerable<DependencyDetail> itemsToUpdate,
+        SourceDependency sourceDependency,
+        IEnumerable<DependencyDetail> oldDependencies,
+        SemanticVersion incomingDotNetSdkVersion);
+
     XmlDocument UpdatePackageSources(XmlDocument nugetConfig, Dictionary<string, HashSet<string>> maestroManagedFeedsByRepo);
 
     Task<bool> Verify(string repo, string branch);
