+++ src/Microsoft.DotNet.Darc/DarcLib/VirtualMonoRepo/PcsVmrForwardFlower.cs
@@ -45,20 +45,16 @@ internal class PcsVmrForwardFlower : VmrForwardFlower, IPcsVmrForwardFlower
         IVmrUpdater vmrUpdater,
         IVmrDependencyTracker dependencyTracker,
         IVmrCloneManager vmrCloneManager,
-        IDependencyFileManager dependencyFileManager,
         IRepositoryCloneManager repositoryCloneManager,
         ILocalGitClient localGitClient,
-        ILocalLibGit2Client libGit2Client,
         IBasicBarClient basicBarClient,
         ILocalGitRepoFactory localGitRepoFactory,
         IVersionDetailsParser versionDetailsParser,
         IProcessManager processManager,
-        ICoherencyUpdateResolver coherencyUpdateResolver,
-        IAssetLocationResolver assetLocationResolver,
         IForwardFlowConflictResolver conflictResolver,
         IFileSystem fileSystem,
         ILogger<VmrCodeFlower> logger)
-        : base(vmrInfo, sourceManifest, vmrUpdater, dependencyTracker, vmrCloneManager, dependencyFileManager, localGitClient, libGit2Client, basicBarClient, localGitRepoFactory, versionDetailsParser, processManager, coherencyUpdateResolver, assetLocationResolver, conflictResolver, fileSystem, logger)
+        : base(vmrInfo, sourceManifest, vmrUpdater, dependencyTracker, vmrCloneManager, localGitClient, basicBarClient, localGitRepoFactory, versionDetailsParser, processManager, conflictResolver, fileSystem, logger)
     {
         _sourceManifest = sourceManifest;
         _dependencyTracker = dependencyTracker;
