+++ src/Microsoft.DotNet.Darc/DarcLib/VirtualMonoRepo/VmrInfo.cs
@@ -83,7 +83,7 @@ public class VmrInfo : IVmrInfo
     public const string CodeownersFileName = "CODEOWNERS";
     public const string CredScanSuppressionsFileName = "CredScanSuppressions.json";
 
-    public static NativePath ArcadeRepoDir = new NativePath(SourceDirName) / "arcade";
+    public static readonly NativePath ArcadeRepoDir = new NativePath(SourceDirName) / "arcade";
 
     public static UnixPath DefaultRelativeSourceMappingsPath { get; } = SourcesDir / SourceMappingsFileName;
